# Execute emulator tasks in main thread