@abejgonzalez 2020-05-27
Right now I am getting 44,657us/22.4fps with a LargeBoom configuration with NVDLA Small. I'm going to confirm that the regressions run on Small.