@abejgonzalez 2020-05-27
I can confirm that this results in a sizeable speedup and works for NVDLA Small (regressions + resnet50) with a BOOM control core.

Before: 479,855us or ~2.1fps
After: 44,657us ~22.4fps